+++ _include/Force_extend.h
@@ -15,13 +15,13 @@ namespace Trajectory
 {
 	//flag=1:lanechangeleft;flag=2:lanechangeright;flag=3:lanechangeleftc;flag=4:lanechangerightc
 	//flag=5:leftturn;flag=6:rightturn;flag=7:leftturnc;flag=8:Uturn
-	bool lanechange(const int &sign, const vector<float> &L_theta, const vector<float> &bound_condition, vector<float> *control);
-	bool lanechange_c(const int &sign, const vector<float> &L_theta, const vector<float> &bound_condition, const vector<float> &constraints, vector<float> *control);
-	bool turn(const int &sign, const vector<float> &L_theta, const vector<float> &bound_condition, vector<float> *control);
-	bool turn_c(const vector<float> &L_theta, const vector<float> &bound_condition, const vector<float> &constraints, vector<float> *control);
-	bool U_turn(const vector<float> &L_theta, const vector<float> &bound_condition, vector<float> *control);
-	bool U_turn_c(const vector<float> &L_theta, const vector<float> &bound_condition, const vector<float> &constraints, vector<float> *control);
-	int is_force_extend_safe(const Vehicle::Node &startnode, const float &step, const vector<float> &control, Collision::collision *collimap, float *le, vector<Vehicle::Node> *path);
+	bool lanechange(const int &sign, const vector<double> &L_theta, const vector<double> &bound_condition, vector<double> *control);
+	bool lanechange_c(const int &sign, const vector<double> &L_theta, const vector<double> &bound_condition, const vector<double> &constraints, vector<double> *control);
+	bool turn(const int &sign, const vector<double> &L_theta, const vector<double> &bound_condition, vector<double> *control);
+	bool turn_c(const vector<double> &L_theta, const vector<double> &bound_condition, const vector<double> &constraints, vector<double> *control);
+	bool U_turn(const vector<double> &L_theta, const vector<double> &bound_condition, vector<double> *control);
+	bool U_turn_c(const vector<double> &L_theta, const vector<double> &bound_condition, const vector<double> &constraints, vector<double> *control);
+	int is_force_extend_safe(const Vehicle::Node &startnode, const double &step, const vector<double> &control, Collision::collision *collimap, double *le, vector<Vehicle::Node> *path);
 }
 
 #endif
