+++ _source/Environment.cpp
@@ -12,9 +12,9 @@ void Environment::EnvironMap::point_construct(const position &point)
 //shape=[length,width]
 void Environment::EnvironMap::vehicle_construct(const position &center, const position &shape)// view as a rectangle 
 {
-	for (float i = center.x; i < center.x + shape.x + interval; i += interval)
+	for (double i = center.x; i < center.x + shape.x + interval; i += interval)
 	{
-		for (float j = center.y - shape.y / 2; j < center.y + shape.y / 2 + interval; j += interval)
+		for (double j = center.y - shape.y / 2; j < center.y + shape.y / 2 + interval; j += interval)
 		{
 			position key(i, j);
 			point_construct(key);
@@ -27,23 +27,23 @@ void Environment::EnvironMap::line_construct(const std::vector<position> &points
 	point_construct(*points.begin());
 	for (auto p = points.begin() + 1; p != points.end(); p++)
 	{
-		float d = sqrtf(pow(p->x - (p - 1)->x, 2) + pow(p->y - (p - 1)->y, 2));
+		double d = sqrt(pow(p->x - (p - 1)->x, 2) + pow(p->y - (p - 1)->y, 2));
 		if (d < interval)
 			point_construct(*p);
 		else
 		{
-			float theta = atan2(p->y - (p - 1)->y, p->x - (p - 1)->x);
+			double theta = atan2(p->y - (p - 1)->y, p->x - (p - 1)->x);
 			int N = std::floor(d / interval);
 			for (int i = 1; i <= N; i++)
-				point_construct(position((p - 1)->x + interval*i*cosf(theta), (p - 1)->y + interval*i*sinf(theta)));
+				point_construct(position((p - 1)->x + interval*i*cos(theta), (p - 1)->y + interval*i*sin(theta)));
 			point_construct(*p);
 		}
 	}
 }
 
-void Environment::EnvironMap::guard_construct(const std::vector<position> &points, const float &width)
+void Environment::EnvironMap::guard_construct(const std::vector<position> &points, const double &width)
 {
-	for (float i = -width / 2; i < width / 2 + interval; i += interval)
+	for (double i = -width / 2; i < width / 2 + interval; i += interval)
 	{
 		std::vector<position> points_width = points;
 		for (auto &p : points_width)
@@ -61,14 +61,14 @@ void Environment::EnvironMap::reset()
 void Environment::map_construct(EnvironMap *environmap)
 {
 	// construct guard 
-	float lane_center = 5.f;// 2.25f;
-	float guard_width = 0.2f;
+	double lane_center = 5.f;// 2.25f;
+	double guard_width = 0.2f;
 	std::vector<position> points;
-	for (float i = 0.f; i <= 4.f; i += environmap->_interval())
+	for (double i = 0.f; i <= 4.f; i += environmap->_interval())
 		points.emplace_back(i, lane_center);
 	environmap->guard_construct(points, guard_width);
 	points.clear();
-	for (float i = 12.f; i <= 33.f; i += environmap->_interval())
+	for (double i = 12.f; i <= 33.f; i += environmap->_interval())
 		points.emplace_back(i, lane_center);
 	environmap->guard_construct(points, guard_width);
 	//// construct obstacle
